add dropdown to after school program form to pick type
